pyproject.toml
[tool.poetry]
name = "search"
version = "0.1.0"
description = ""
authors = ["kalio007 <kalioprince1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
elasticsearch = "^8.13.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
